Refactored AnalysisJobBuilderWindowImpl to have "classic view" in separate class - JobClassicView
